# frontend/openapi.json
{
  "components": {
    "schemas": {
      "BenchmarkManifest": {
        "additionalProperties": false,
        "properties": {
          "completed_cells": {
            "default": 0,
            "title": "Completed Cells",
            "type": "integer"
          },
          "created_at_utc": {
            "format": "date-time",
            "title": "Created At Utc",
            "type": "string"
          },
          "dataset": {
            "title": "Dataset",
            "type": "string"
          },
          "dataset_revision": {
            "title": "Dataset Revision",
            "type": "string"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "failed_cells": {
            "default": 0,
            "title": "Failed Cells",
            "type": "integer"
          },
          "generation_config": {
            "additionalProperties": true,
            "title": "Generation Config",
            "type": "object"
          },
          "generation_config_hash": {
            "title": "Generation Config Hash",
            "type": "string"
          },
          "git_commit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Git Commit"
          },
          "mode": {
            "$ref": "#/components/schemas/BenchmarkMode"
          },
          "model_ids": {
            "items": {
              "type": "string"
            },
            "title": "Model Ids",
            "type": "array"
          },
          "registry_models_sha256": {
            "title": "Registry Models Sha256",
            "type": "string"
          },
          "registry_snapshot_id": {
            "title": "Registry Snapshot Id",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "schema_version": {
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/BenchmarkStatus"
          },
          "successful_cells": {
            "default": 0,
            "title": "Successful Cells",
            "type": "integer"
          },
          "total_cells": {
            "title": "Total Cells",
            "type": "integer"
          },
          "updated_at_utc": {
            "format": "date-time",
            "title": "Updated At Utc",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "mode",
          "status",
          "dataset",
          "dataset_revision",
          "registry_snapshot_id",
          "registry_models_sha256",
          "model_ids",
          "generation_config",
          "generation_config_hash",
          "git_commit",
          "total_cells",
          "created_at_utc",
          "updated_at_utc"
        ],
        "title": "BenchmarkManifest",
        "type": "object"
      },
      "BenchmarkMode": {
        "enum": [
          "pilot",
          "full"
        ],
        "title": "BenchmarkMode",
        "type": "string"
      },
      "BenchmarkRecord": {
        "additionalProperties": false,
        "properties": {
          "configuration": {
            "title": "Configuration",
            "type": "string"
          },
          "dataset": {
            "title": "Dataset",
            "type": "string"
          },
          "exact_match": {
            "title": "Exact Match",
            "type": "integer"
          },
          "expected_answer": {
            "title": "Expected Answer",
            "type": "string"
          },
          "extracted_answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extracted Answer"
          },
          "extraction_method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Method"
          },
          "generation_config_hash": {
            "title": "Generation Config Hash",
            "type": "string"
          },
          "mode": {
            "$ref": "#/components/schemas/BenchmarkMode"
          },
          "model_snapshot_id": {
            "title": "Model Snapshot Id",
            "type": "string"
          },
          "prompt_template_version": {
            "title": "Prompt Template Version",
            "type": "string"
          },
          "rendered_prompt": {
            "title": "Rendered Prompt",
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/GenerationResult"
          },
          "revision": {
            "title": "Revision",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "sample_id": {
            "title": "Sample Id",
            "type": "string"
          },
          "schema_version": {
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "source_index": {
            "title": "Source Index",
            "type": "integer"
          },
          "split": {
            "title": "Split",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "mode",
          "dataset",
          "configuration",
          "revision",
          "split",
          "source_index",
          "sample_id",
          "prompt_template_version",
          "rendered_prompt",
          "expected_answer",
          "extracted_answer",
          "extraction_method",
          "exact_match",
          "result",
          "model_snapshot_id",
          "generation_config_hash"
        ],
        "title": "BenchmarkRecord",
        "type": "object"
      },
      "BenchmarkRunView": {
        "additionalProperties": false,
        "properties": {
          "manifest": {
            "$ref": "#/components/schemas/BenchmarkManifest"
          },
          "summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BenchmarkSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "manifest"
        ],
        "title": "BenchmarkRunView",
        "type": "object"
      },
      "BenchmarkStartRequest": {
        "additionalProperties": false,
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/BenchmarkMode",
            "default": "full"
          }
        },
        "title": "BenchmarkStartRequest",
        "type": "object"
      },
...
      "BenchmarkSummary": {
        "additionalProperties": false,
        "properties": {
          "generated_at_utc": {
            "format": "date-time",
            "title": "Generated At Utc",
            "type": "string"
          },
          "models": {
            "items": {
              "$ref": "#/components/schemas/ModelBenchmarkSummary"
            },
            "title": "Models",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "generated_at_utc",
          "models"
        ],
        "title": "BenchmarkSummary",
        "type": "object"
      },
...
      "ComparisonRequest": {
        "additionalProperties": false,
        "properties": {
          "prompt": {
            "maxLength": 50000,
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          },
          "save": {
            "default": false,
            "title": "Save",
            "type": "boolean"
          }
        },
        "required": [
          "prompt"
        ],
        "title": "ComparisonRequest",
        "type": "object"
      },
      "ComparisonResponse": {
        "additionalProperties": false,
        "properties": {
          "comparison_id": {
            "format": "uuid",
            "title": "Comparison Id",
            "type": "string"
          },
          "created_at_utc": {
            "format": "date-time",
            "title": "Created At Utc",
            "type": "string"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/GenerationResult"
            },
            "title": "Results",
            "type": "array"
          },
          "saved": {
            "title": "Saved",
            "type": "boolean"
          }
        },
        "required": [
          "prompt",
          "saved",
          "results"
        ],
        "title": "ComparisonResponse",
        "type": "object"
      },
...
      "EligibilityRequest": {
        "additionalProperties": false,
        "properties": {
          "min_context_tokens": {
            "default": 500000,
            "minimum": 1.0,
            "title": "Min Context Tokens",
            "type": "integer"
          }
        },
        "title": "EligibilityRequest",
        "type": "object"
      },
      "EligibilityResponse": {
        "additionalProperties": false,
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/EligibilityResult"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "results"
        ],
        "title": "EligibilityResponse",
        "type": "object"
      },
      "EligibilityResult": {
        "additionalProperties": false,
        "properties": {
          "eligible": {
            "title": "Eligible",
            "type": "boolean"
          },
          "evaluated_requirements": {
            "$ref": "#/components/schemas/EligibilityRequest"
          },
          "model_id": {
            "title": "Model Id",
            "type": "string"
          },
          "model_snapshot_id": {
            "title": "Model Snapshot Id",
            "type": "string"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          }
        },
        "required": [
          "model_id",
          "eligible",
          "reasons",
          "evaluated_requirements",
          "model_snapshot_id"
        ],
        "title": "EligibilityResult",
        "type": "object"
      },
      "GenerationResult": {
        "additionalProperties": false,
        "properties": {
          "attempt_number": {
            "maximum": 2.0,
            "minimum": 1.0,
            "title": "Attempt Number",
            "type": "integer"
          },
          "completed_at_utc": {
            "format": "date-time",
            "title": "Completed At Utc",
            "type": "string"
          },
          "completion_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completion Tokens"
          },
          "cost_source": {
            "$ref": "#/components/schemas/CostSource",
            "default": "unavailable"
          },
          "cost_usd": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Usd"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "latency_ms": {
            "minimum": 0.0,
            "title": "Latency Ms",
            "type": "number"
          },
          "prompt_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Tokens"
          },
          "provider": {
            "default": "openrouter",
            "title": "Provider",
            "type": "string"
          },
          "provider_request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Request Id"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "requested_model_id": {
            "title": "Requested Model Id",
            "type": "string"
          },
          "served_model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Served Model Id"
          },
          "started_at_utc": {
            "format": "date-time",
            "title": "Started At Utc",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/GenerationStatus"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "total_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Tokens"
          }
        },
        "required": [
          "request_id",
          "requested_model_id",
          "status",
          "latency_ms",
          "started_at_utc",
          "completed_at_utc",
          "attempt_number"
        ],
        "title": "GenerationResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PaginatedRecords": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BenchmarkRecord"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "PaginatedRecords",
        "type": "object"
      },
...
      "RouteRequest": {
        "additionalProperties": false,
        "properties": {
          "min_context_tokens": {
            "default": 1,
            "minimum": 1.0,
            "title": "Min Context Tokens",
            "type": "integer"
          },
          "priority": {
            "$ref": "#/components/schemas/RoutingPriority",
            "default": "balanced"
          },
          "prompt": {
            "maxLength": 50000,
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          }
        },
        "required": [
          "prompt"
        ],
        "title": "RouteRequest",
        "type": "object"
      },
      "RouteResponse": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/RoutingDecision"
          },
          "result": {
            "$ref": "#/components/schemas/GenerationResult"
          }
        },
        "required": [
          "decision",
          "result"
        ],
        "title": "RouteResponse",
        "type": "object"
      },
...
      "RoutingPriority": {
        "enum": [
          "cost",
          "quality",
          "latency",
          "balanced"
        ],
        "title": "RoutingPriority",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Checkpoint 2 transparent routing over a reproducible model benchmark API",
    "title": "Adaptive LLM Router",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/benchmarks/gsm8k/runs": {
      "post": {
        "operationId": "start_benchmark_api_v1_benchmarks_gsm8k_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BenchmarkStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkRunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Benchmark"
      }
    },
...
    "/api/v1/benchmarks/runs/{run_id}": {
      "get": {
        "operationId": "benchmark_run_api_v1_benchmarks_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkRunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Benchmark Run"
      }
    },
    "/api/v1/benchmarks/runs/{run_id}/records": {
      "get": {
        "operationId": "benchmark_records_api_v1_benchmarks_runs__run_id__records_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedRecords"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Benchmark Records"
      }
    },
    "/api/v1/comparisons": {
      "post": {
        "operationId": "compare_api_v1_comparisons_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComparisonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComparisonResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compare"
      }
    },
    "/api/v1/eligibility": {
      "post": {
        "operationId": "eligibility_api_v1_eligibility_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EligibilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EligibilityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Eligibility"
      }
    },
...
    "/api/v1/routes": {
      "post": {
        "description": "Select exactly one pinned model using the explicit CP2 rule policy.",
        "operationId": "route_prompt_api_v1_routes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Route Prompt"
      }
    },
...
  }
}